move signal management to utils and strategy selection to optimizer
